+++ Makefile
@@ -3,6 +3,6 @@ all: build test
 test:
 	$(GO_CMD) test -v ./...
 build:
-	$(GO_CMD) build -o main ./main.go
+	$(GO_CMD) build -o main ./$(service)/cmd/main.go
 clean:
 	rm -f main
