+++ auth-service/internal/internal.go
@@ -1,0 +1,38 @@
+package internal
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/karlo235/DHBW-DevOps/auth-service/pkg"
+)
+
+func AuthLoginHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+
+	username := r.FormValue("username")
+	password := r.FormValue("password")
+
+	// For simplicity, we'll use a hardcoded username and password
+	// This should be replaced with a proper authentication mechanism
+	if username == "user" && password == "pass" {
+		token, err := pkg.CreateToken(username)
+
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			w.Write([]byte(`{"error": "Error generating the token"}`))
+			return
+		}
+
+		w.Write([]byte(fmt.Sprintf(`{"token": "%s"}`, token)))
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+		w.Write([]byte(`{"error": "Invalid credentials"}`))
+	}
+}
+
+func AuthLogoutHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	// In this simple example, we'll just return a success message
+	w.Write([]byte(`{"message": "Logout successful"}`))
+}
