+++ checkout-service/internal/internal.go
@@ -1,0 +1,30 @@
+package internal
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/gorilla/mux"
+	"github.com/karlo235/DHBW-DevOps/checkout-service/pkg"
+)
+
+func CheckoutPlaceOrderHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	token := r.Header.Get("Authorization")
+
+	if token == "" {
+		w.WriteHeader(http.StatusUnauthorized)
+		w.Write([]byte(`{"error": "Missing authorization header"}`))
+		return
+	}
+
+	if pkg.VerifyToken(token) {
+		// In this simple example, we'll just return a success message
+		w.Write([]byte(`{"message": "Order placed successfully"}`))
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+		w.Write([]byte(`{"error": "Invalid token"}`))
+	}
+}
