+++ product-service/internal/internal.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 
 	"github.com/gorilla/mux"
-	"github.com/karlo235/DHBW-DevOps/pkg"
+	"github.com/karlo235/DHBW-DevOps/product-service/pkg"
 )
 
 type Product struct {
@@ -21,36 +21,6 @@ var products = []Product{
 	{ID: 1, Name: "Pants", Price: 19.99},
 	{ID: 2, Name: "Dress", Price: 29.99},
 	{ID: 3, Name: "Shirt", Price: 39.99},
-}
-
-func AuthLoginHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	username := r.FormValue("username")
-	password := r.FormValue("password")
-
-	// For simplicity, we'll use a hardcoded username and password
-	// This should be replaced with a proper authentication mechanism
-	if username == "user" && password == "pass" {
-		token, err := pkg.CreateToken(username)
-
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte(`{"error": "Error generating the token"}`))
-			return
-		}
-
-		w.Write([]byte(fmt.Sprintf(`{"token": "%s"}`, token)))
-	} else {
-		w.WriteHeader(http.StatusUnauthorized)
-		w.Write([]byte(`{"error": "Invalid credentials"}`))
-	}
-}
-
-func AuthLogoutHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	// In this simple example, we'll just return a success message
-	w.Write([]byte(`{"message": "Logout successful"}`))
 }
 
 func FindProductByID(products []Product, id int) *Product {
@@ -106,23 +76,4 @@ func ProductDetailHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.Write(response)
-}
-
-func CheckoutPlaceOrderHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	token := r.Header.Get("Authorization")
-
-	if token == "" {
-		w.WriteHeader(http.StatusUnauthorized)
-		w.Write([]byte(`{"error": "Missing authorization header"}`))
-		return
-	}
-
-	if pkg.VerifyToken(token) {
-		// In this simple example, we'll just return a success message
-		w.Write([]byte(`{"message": "Order placed successfully"}`))
-	} else {
-		w.WriteHeader(http.StatusUnauthorized)
-		w.Write([]byte(`{"error": "Invalid token"}`))
-	}
 }
